requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.7
aiohttp==3.9.1
lxml==4.9.3
python-dotenv==1.0.0
